[기능추가] : 댓글 등록 기능 추가(back) / Front : 405 에러

+++ src/main/java/com/insilicogen/CRUD_PRJ/user/service/User.java
@@ -28,7 +28,7 @@ import lombok.Setter;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userSn;
+    private Long userSn;
 
     @Column(name = "USER_LGN_ID", length = 30)
     @NotNull
